New Batch of best bots
The Cost bot trys to predict cost of doing each action 100 times in a row, then makes a decision on that. The dec bot simply tries to predict the best decision, with the same metric, but has no concept of cost magnitude. test_cost and test_bot run validation on the cost and dec bots respectively, since they have different architectures you have to use the correct testing bot.