Plotting During Learning
This is the bot that works best, using the 100 steps in each direction ground truth method, with a running plot (and saved) plot of the error during each epoch. In my opinion, the spikes at 5, 7.5, 9, 11, thousand steps confirm the level theory, along with the T-SNE clustering. If only I had a faster computer.